Category fix
If it is a non spend transaction and we find the merchant  - we need to use subtype category - more than merchant category. So subtype value takes a preference for category testing.

Example: Walmart Payroll -- merchant is walmart, subtype will be direct deposit and category should be paychecks/Salary

without this change it would show up as General Merchandise when it finds walmart - regardless of transaction subtype.